+++ docs/changelog.md
@@ -8,11 +8,13 @@
 - Added new languages
   - Azerbaijani
   - Slovak
+- Added search filters
 
 ### Fix bugs
 
 - Fixed error `NameError: name 'exit' is not defined` in bundled executable 
 when executing command `login_cache` or `login_cache:show`
+- Fixed mangadex-downloader won't start in Python 3.8
 
 ### Breaking changes